feat: resize images fix: the images are now remove from the file images when deleted

+++ Controllers/Book.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const path = require('path');
 const Book = require('../models/Books');
 
 exports.addBook = (req, res) => {
@@ -40,14 +41,19 @@ exports.deleteBook = (req, res) => {
   Book.findOne({ _id: req.params.id })
     .then((book) => {
       if (book.userId !== req.auth.userId) {
-        res.status(401).json({ message: 'Not authorized' });
-      } else {
+        return res.status(401).json({ message: 'Not authorized' });
+      }
+      if (book.imageUrl) {
         const filename = book.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-          Book.deleteOne({ _id: req.params.id })
-            .then(() => { res.status(200).json({ message: 'Objet supprimé !' }); })
-            .catch((error) => res.status(401).json({ error }));
+        const filePath = path.join(__dirname, '..', 'public', 'images', filename);
+        fs.unlink(filePath, (err) => {
+          if (err) {
+            return res.status(500).json({ error: 'File deletion failed' });
+          }
         });
+        Book.deleteOne({ _id: req.params.id })
+          .then(() => { res.status(200).json({ message: 'Objet supprimé !' }); })
+          .catch((error) => res.status(401).json({ error }));
       }
     })
     .catch((error) => {
